refactor: chunk into files
